Separate controller/reconcile logic
This ensures that between controller definition
and reconcile logic, we have a clear separation.

This also add a missing suite for the resources internal pkg.
