+++ pkg/reconciler/buildrun/resources/resources_suite_test.go
@@ -1,0 +1,13 @@
+package resources_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestResources(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Resources Suite")
+}
